api/requirements.txt
fastapi 
boto3
uvicorn
python-dotenv